+++ Editor/Framework/Watch/WatchPropertyDrawer.cs
@@ -1,3 +1,5 @@
+#if UNITY_EDITOR
+
 using UnityEditor;
 using UnityEngine;
 
@@ -13,3 +15,5 @@ namespace DT.UniStart.Editor {
     }
   }
 }
+
+#endif